Check existence of expression in console endpoint
Previously, using the `/api/user/console` endpoint and specifying a shard
but not an expression resulted in the user being sent a `TypeError`
description in the response JSON. Now, the user is told that the
expression is invalid.

The error expression was chosen to be similar to that for an invalid
shard.